chore: update spec-update.group-imports
