+++ tests/spec/failed-tests.txt
@@ -1,5 +1,3 @@
-FAIL  tests/spec/spec-update.group-imports.spec.ts
-
 FAIL  tests/spec/all.import-multiple.spec.ts
 
 FAIL  tests/update-custom/update-custom-green-path.debug-logs.spec.ts
