+++ tests/spec/test-data/spec-update.group-imports/group-imports-from-dom.spec.ts
@@ -1,0 +1,2 @@
+describe('C', () => {
+});
