+++ tests/spec/test-data/spec-update.group-imports/group-imports-from-dom.ts
@@ -1,0 +1,7 @@
+export class C  {
+    constructor(
+        private aDep: Event,
+        private bDep: Object,
+        private cDep: Window,
+    ) {}
+}
