+++ tests/spec/test-data/spec-update.group-imports/group-imports.spec.ts
@@ -1,0 +1,6 @@
+import { BDep, LogService } from '@angular/core';
+import { other } from 'other';
+import { some } from '../util';
+
+describe('C', () => {
+});
